Added some typedefs and structs copied from pcb.
Added all typedefs and structs for holding data for an element.
Added doxygen type comments to clarify some fields in the structs.
This way either a preview or a footprint file can be generated
from the same set of variables.
In the future this may simplify the conversion of pcb-gfpw from a
stand-alone application to a pcb plug-in.