Added a function registering feature for package type dependant functions.
This mechanism is not operational yet, just added for future usage.
This mechanism is copied from the pcb plug-in/HID mechanism.
The usual copyright disclaimer applies here and my thanks go to
DJ Delorie for creating this functionality in pcb.
This implementation is modified for the specific needs of pcb-gfpw/fpw.