Create a header for doxygen docs with git describe tag.
I added two shell scripts for generating doxygen developer docs
with and without callgraphs.
The scripts first generate a html header file and insert the git describe
tag with the current tag, the number of commits after this tag and the first
8 characters of the hash of the most recent commit (HEAD).
When finished with header the developer docs are generated by calling
doxygen with the configuration file of choice.