Correction on the pin 1 position and rotated 90 degrees CW [#27].
Parts with pins/pads on two sideshave the pin 1 position
defined as: left side, top pin.
This resulted in a rotation of 90 degrees CW to meet this
definition.
As the pin 1 marker was drawn outside the courtyard this
is to be corrected in a following commit.
Code for creating this piece of silk art is now removed.