Added appdata.
Many distributions are now shipping the gnome-software tool,
which gives user-friendly information about applications in
a sort of "app store" format.

We should ship a geda-gaf.appdata.xml file in our tarball
so that the the gEDA tools appear in the same way to users
no matter which distribution they happen to be using.

Signed-off-by: Bert Timmerman <bert.timmerman@xs4all.nl>